requirements.txt
django-debug-toolbar
pycbrf @ git+https://github.com/idlesign/pycbrf.git
tinkoff-investments
